Added code syntax highlighting, page footer, and moved prev\next links
Added code syntax highlighting with [Prism](http://prismjs.com) and the Atom Dark theme for
Prisim. Modified `haunt.css` to provide some style overrides, as well as a fallback dark
theme if javascript is disabled or doesn't load. Also added footer section (through a partial)
and moved the previous\next post from the bottom of the post pages to the sidebar.